requirements.txt
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
transformers==4.25.1
datasets==2.8.0
scipy==1.10.0
librosa==0.9.2
jiwer==2.5.1
evaluate==0.4.0
click==8.1.3
pandas==1.5.3
praat-parselmouth==1.23.5
mutagen==1.46.0
azure-cognitiveservices-speech==1.25.0